# devspace-chat-backend/src/main/java/com/stmoxx/devspacechat/common/StringResponse.java
package com.stmoxx.devspacechat.common;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class StringResponse {

    private String response;
}